OPENEUROPA-1972: Add another plugin base for non ief content forms.
